# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tech Mentor ERP & POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # PRODUCTS
...
  /subcategories:
    get:
      operationId: listSubcategories
      tags: [subcategories]
      parameters:
        - { name: categoryId, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: List of subcategories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subcategory"
    post:
      operationId: createSubcategory
      tags: [subcategories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
  /subcategories/{id}:
    patch:
      operationId: updateSubcategory
      tags: [subcategories]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
    delete:
      operationId: deleteSubcategory
      tags: [subcategories]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # BRANDS
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      responses:
        "200":
          description: List of routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeliveryRoute"
    post:
      operationId: createRoute
      tags: [routes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryRouteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryRoute"
  /routes/{id}:
    patch:
      operationId: updateRoute
      tags: [routes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryRouteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryRoute"
    delete:
      operationId: deleteRoute
      tags: [routes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # DELIVERY ASSIGNMENTS
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
...
    User:
      type: object
      required: [id, username, role, name, status, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        email: { type: ["string", "null"] }
        role: { type: string }
        status: { type: string }
        branchId: { type: ["integer", "null"] }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [username, password, name, role]
      properties:
        username: { type: string }
        password: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        branchId: { type: integer }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        role: { type: string }
        status: { type: string }
        branchId: { type: integer }

    # DASHBOARD
...
    Subcategory:
      type: object
      required: [id, name, categoryId, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        categoryId: { type: integer }
        categoryName: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        productCount: { type: integer }
        createdAt: { type: string }
    SubcategoryInput:
      type: object
      required: [name, categoryId]
      properties:
        name: { type: string }
        categoryId: { type: integer }
        description: { type: string }
    SubcategoryUpdate:
      type: object
      properties:
        name: { type: string }
        categoryId: { type: integer }
        description: { type: string }

    # BRAND
...
    DeliveryRoute:
      type: object
      required: [id, name, area, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        area: { type: string }
        vehicle: { type: ["string", "null"] }
        employeeId: { type: ["integer", "null"] }
        employeeName: { type: ["string", "null"] }
        deliveryDate: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
    DeliveryRouteInput:
      type: object
      required: [name, area]
      properties:
        name: { type: string }
        area: { type: string }
        vehicle: { type: string }
        employeeId: { type: integer }
        deliveryDate: { type: string }
    DeliveryRouteUpdate:
      type: object
      properties:
        name: { type: string }
        area: { type: string }
        vehicle: { type: string }
        employeeId: { type: integer }
        deliveryDate: { type: string }
        status: { type: string }

    # DELIVERY ASSIGNMENT
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string